added separate logging for hyperparameter optimization procedure.

+++ flair/training_utils.py
@@ -1,6 +1,7 @@
 import itertools
 import random
 import logging
+import os.path
 from collections import defaultdict
 from enum import Enum
 from math import inf
@@ -483,7 +484,8 @@ def init_output_file(base_path: Union[str, Path], file_name: str) -> Path:
     base_path.mkdir(parents=True, exist_ok=True)
 
     file = base_path / file_name
-    open(file, "w", encoding="utf-8").close()
+    if not os.path.exists(file):
+        open(file, "w", encoding="utf-8").close()
     return file
 
 
@@ -506,9 +508,9 @@ def log_line(log):
     log.info("-" * 100)
 
 
-def add_file_handler(log, output_file):
+def add_file_handler(log, output_file, mode="w", initialize=False):
     init_output_file(output_file.parents[0], output_file.name)
-    fh = logging.FileHandler(output_file, mode="w", encoding="utf-8")
+    fh = logging.FileHandler(output_file, mode=mode, encoding="utf-8")
     fh.setLevel(logging.INFO)
     formatter = logging.Formatter("%(asctime)-15s %(message)s")
     fh.setFormatter(formatter)
